Libraries shared types #9726
Renamed `shared` to `core`.
Optimized build.
Fixed typos.
