+++ modules/lib/core/index.d.ts
@@ -1,10 +1,13 @@
 declare global {
-    // eslint-disable-next-line @typescript-eslint/no-empty-interface
     interface XpXData {
     }
 }
 
-import {Component, Region} from './portal';
+export type UserKey = `user:${string}:${string}`;
+export type GroupKey = `group:${string}:${string}`;
+export type RoleKey = `role:${string}`;
+
+export type PrincipalKey = UserKey | GroupKey | RoleKey;
 
 export type WorkflowState = 'IN_PROGRESS' | 'PENDING_APPROVAL' | 'REJECTED' | 'READY';
 
@@ -28,6 +31,19 @@ export interface PublishInfo {
     from?: string;
     to?: string;
     first?: string;
+}
+
+export interface Component<Config extends object = object, Regions extends Record<string, Region> = Record<string, Region>> {
+    config: Config;
+    descriptor: string;
+    path: string;
+    type: 'page' | 'layout' | 'part';
+    regions: Regions;
+}
+
+export interface Region<Config extends object = object> {
+    name: string;
+    components: Component<Config>[];
 }
 
 export interface Content<
