--- modules/lib/shared/auth.d.ts
@@ -1,5 +1,0 @@
-export type UserKey = `user:${string}:${string}`;
-export type GroupKey = `group:${string}:${string}`;
-export type RoleKey = `role:${string}`;
-
-export type PrincipalKey = UserKey | GroupKey | RoleKey;
