--- modules/lib/shared/index.d.ts
@@ -1,4 +1,0 @@
-export * from './auth';
-export * from './content';
-export * from './portal';
-export * from './utils';
