--- modules/lib/shared/portal.d.ts
@@ -1,12 +1,0 @@
-export interface Component<Config extends object = object, Regions extends Record<string, Region> = Record<string, Region>> {
-    config: Config;
-    descriptor: string;
-    path: string;
-    type: 'page' | 'layout' | 'part';
-    regions: Regions;
-}
-
-export interface Region<Config extends object = object> {
-    name: string;
-    components: Component<Config>[];
-}
